requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
PyJWT
cryptography
